[TERK-92/Fix] 정산 내역 조회 및 수정 페이지 변경사항 반영
