+++ src/components/cartsharecal/create/CalCreateOptInp.js
@@ -1,75 +1,61 @@
 /** @jsxImportSource @emotion/react */
 import styled from 'styled-components';
-import NavigationBar from '../cartshare/NavigationBar';
-import DutchPayNavigationBar from './CartShareCalNavigationBar';
-import Master from './Master';
-import Me from './Me';
+import NameInfoCreate from './NameInfoCreate';
 
-const DutchPayUpdateOptInp = ({ dutchPay, inpInput, onChangeInpInput, onChangeInpRmd }) => {
+const CalCreateOptInp = ({ cartShareCal, inpInput, onChangeInpInput, onChangeInpRmd }) => {
     return (
-        <OptSecContainer>
-            {dutchPay.dutchPayDtlList?.map((info, index) => (
-                <DutchPayDtl key={info.mbrId}>
-                    <div className="name-container">
-                        {info.mastrYn && <Me />}
-                        <div className="name">
-                            {info.mbrNm}
-                            {info.mastrYn && <Master />}
-                        </div>
-                    </div>
-                    <div className="dt-container">
-                        <div className="main-dt-container">
+        <OptInpContainer>
+            {cartShareCal.cartShareCalDtlList?.map((info, index) => (
+                <CalDtl key={info.mbrId}>
+                    <CalDtlInfo>
+                        <NameInfoCreate info={info} />
+
+                        <div className="main-cal">
                             <input
-                                className="dt-value"
+                                className="main-cal-amt"
                                 onChange={e => onChangeInpInput(e, info.mbrId)}
-                                value={Number(inpInput.dtlMap.get(info.mbrId)).toString()}
-                                type="number"
+                                value={inpInput.dtlMap.get(info.mbrId).toLocaleString()}
                             />
-                            <div className="dt-unit">원</div>
+                            <div className="main-cal-unit">원</div>
                         </div>
-                    </div>
-                </DutchPayDtl>
+                    </CalDtlInfo>
+                </CalDtl>
             ))}
             <div className="rmd">
                 <div className="rmd-label">나머지</div>
                 <div className="rmd-value">
-                    <input
-                        className="rmd-amt"
-                        onChange={onChangeInpRmd}
-                        value={Number(inpInput.dtRmd).toString()}
-                        type="number"
-                    />
+                    <input className="rmd-amt" onChange={onChangeInpRmd} value={inpInput.calRmd.toLocaleString()} />
                     <div className="rmd-unit">원</div>
                 </div>
             </div>
             <div className="border"></div>
-            <div className="sum-container">
+            <div className="sum">
                 <div className="sum-label">총 정산 금액</div>
                 <div className="sum-value">
-                    <div className="sum-amt">{inpInput.dtAmt}</div>
+                    <div className="sum-amt">{inpInput.calAmt.toLocaleString()}</div>
                     <div className="sum-unit">원</div>
                 </div>
             </div>
-            <div className="paymt-container">
+            <div className="paymt">
                 <div className="paymt-label">결제 금액</div>
                 <div className="paymt-value">
-                    <div className="paymt-amt">{dutchPay.paymtAmt}</div>
+                    <div className="paymt-amt">{cartShareCal.ttlPaymtAmt.toLocaleString()}</div>
                     <div className="paymt-unit">원</div>
                 </div>
             </div>
-        </OptSecContainer>
+        </OptInpContainer>
     );
 };
 
-const OptSecContainer = styled.div`
+const OptInpContainer = styled.div`
     .rmd {
         margin-top: 32px;
         display: flex;
         justify-content: space-between;
         font-weight: 400;
         font-size: 16px;
-        height: 15px;
-        line-height: 15px;
+        height: 20px;
+        line-height: 20px;
     }
     .rmd-label {
         padding-left: 46px;
@@ -81,16 +67,16 @@ const OptSecContainer = styled.div`
     }
 
     .rmd-amt {
-        width: 15px;
+        width: 20px;
         text-align: right;
         border: none;
         border-bottom: 1px solid black;
-        padding-right: 4px;
         font-size: 16px;
-        height: 15px;
-        line-height: 15px;
         font-family: 'line';
+        box-sizing: border-box;
+        margin-right: 3px;
     }
+
     .rmd-amt:focus {
         outline: none;
     }
@@ -100,7 +86,7 @@ const OptSecContainer = styled.div`
         border-bottom: 1px solid #888888;
         width: 327px;
     }
-    .sum-container {
+    .sum {
         display: flex;
         margin-top: 26px;
         justify-content: space-between;
@@ -125,15 +111,14 @@ const OptSecContainer = styled.div`
         font-weight: 700;
         font-size: 24px;
         border: none;
-        width: 80px;
+        width: 100px;
         text-align: right;
-        height: 24px;
         padding-right: 4px;
     }
 
-    .paymt-container {
+    .paymt {
         display: flex;
-        margin-top: 10px;
+        margin-top: 15px;
         justify-content: space-between;
         font-weight: 400;
         font-size: 14px;
@@ -154,7 +139,7 @@ const OptSecContainer = styled.div`
 
     .paymt-unit {
         margin-right: 20px;
-        margin-left: 13px;
+        margin-left: 14px;
     }
     input::-webkit-outer-spin-button,
     input::-webkit-inner-spin-button {
@@ -167,52 +152,41 @@ const OptSecContainer = styled.div`
         -moz-appearance: textfield;
     }
 `;
-const DutchPayDtl = styled.div`
+const CalDtl = styled.div`
     margin-top: 20px;
     margin-bottom: 50px;
+`;
+
+const CalDtlInfo = styled.div`
     display: flex;
     justify-content: space-between;
-
-    .name-container {
+    .main-cal {
         display: flex;
-        position: relative;
-    }
-    .name {
         font-weight: 700;
         font-size: 20px;
-        position: absolute;
-        left: 33px;
-        display: flex;
+        margin-bottom: 8px;
+        height: 25px;
+        line-height: 25px;
     }
-    .dt-container {
-        position: relative;
-    }
-    .dt-value {
+
+    .main-cal-amt {
         border: none;
         border-bottom: 1px solid black;
         font-size: 20px;
         font-weight: 700;
         width: 100px;
         text-align: right;
-        height: 25px;
-        line-height: 25px;
-        padding-right: 4px;
         font-family: 'line';
+        box-sizing: border-box;
     }
-    .dt-value:focus {
+    .main-cal-amt:focus {
         outline: none;
     }
 
-    .main-dt-container {
-        display: flex;
-        font-weight: 700;
-        font-size: 20px;
-        margin-bottom: 8px;
-    }
-    .dt-unit {
+    .main-cal-unit {
         margin-right: 14px;
         margin-left: 10px;
     }
 `;
 
-export default DutchPayUpdateOptInp;
+export default CalCreateOptInp;
