+++ src/components/cartsharecal/detail/CalDetailOptInp.js
@@ -1,70 +1,57 @@
 /** @jsxImportSource @emotion/react */
 import styled from 'styled-components';
-import NavigationBar from '../cartshare/NavigationBar';
-import DutchPayNavigationBar from './CartShareCalNavigationBar';
-import Master from './Master';
-import Me from './Me';
+import NameInfoDetail from './NameInfoDetail';
 
-const DutchPayUpdateOptSpl = ({ dutchPay, splInput, onChangeSplInput }) => {
+const CalDetailOptInp = ({ cartShareCal, onClickCmplYn }) => {
     return (
-        <OptSecContainer>
-            {dutchPay.dutchPayDtlList?.map((info, index) => (
-                <DutchPayDtl key={info.mbrId}>
-                    <div className="name-container">
-                        {info.mastrYn && <Me />}
-                        <div className="name">
-                            {info.mbrNm}
-                            {info.mastrYn && <Master />}
-                        </div>
-                    </div>
+        <OptInpContainer>
+            {cartShareCal.cartShareCalDtlList?.map((info, index) => (
+                <CalDtl key={info.mbrId}>
+                    <NameInfoDetail cartShareCal={cartShareCal} info={info} onClickCmplYn={onClickCmplYn} />
                     <div className="dt-container">
                         <div className="main-dt-container">
-                            <div className="dt-value">{splInput.dtDtlAmt}</div>
+                            <div className="dt-value">{info.calDtlAmt.toLocaleString()}</div>
                             <div className="dt-unit">원</div>
                         </div>
                     </div>
-                </DutchPayDtl>
+                </CalDtl>
             ))}
             <div className="rmd">
                 <div className="rmd-label">나머지</div>
                 <div className="rmd-value">
-                    <div className="rmd-amt">{splInput.dtRmd}</div>
+                    <div className="rmd-amt">{cartShareCal.calRmd.toLocaleString()}</div>
                     <div className="rmd-unit">원</div>
                 </div>
             </div>
             <div className="border"></div>
-            <div className="sum-container">
+            <div className="sum">
                 <div className="sum-label">총 정산 금액</div>
                 <div className="sum-value">
-                    <input
-                        className="sum-amt"
-                        onChange={onChangeSplInput}
-                        value={Number(splInput.dtAmt).toString()}
-                        type="number"
-                    />
+                    <div className="sum-amt">{cartShareCal.calAmt.toLocaleString()}</div>
+
                     <div className="sum-unit">원</div>
                 </div>
             </div>
-            <div className="paymt-container">
+            <div className="paymt">
                 <div className="paymt-label">결제 금액</div>
                 <div className="paymt-value">
-                    <div className="paymt-amt">{dutchPay.paymtAmt}</div>
+                    <div className="paymt-amt">{cartShareCal.ttlPaymtAmt.toLocaleString()}</div>
                     <div className="paymt-unit">원</div>
                 </div>
             </div>
-        </OptSecContainer>
+        </OptInpContainer>
     );
 };
 
-const OptSecContainer = styled.div`
+const OptInpContainer = styled.div`
     .rmd {
         margin-top: 32px;
         display: flex;
         justify-content: space-between;
         font-weight: 400;
         font-size: 16px;
-        height: 15px;
-        line-height: 15px;
+        height: 20px;
+        line-height: 20px;
     }
     .rmd-label {
         padding-left: 46px;
@@ -76,15 +63,16 @@ const OptSecContainer = styled.div`
     }
 
     .rmd-amt {
-        margin-right: 3px;
+        padding-right: 3px;
     }
+
     .border {
         margin: auto;
         margin-top: 24px;
         border-bottom: 1px solid #888888;
         width: 327px;
     }
-    .sum-container {
+    .sum {
         display: flex;
         margin-top: 26px;
         justify-content: space-between;
@@ -112,21 +100,20 @@ const OptSecContainer = styled.div`
         width: 80px;
         text-align: right;
         height: 24px;
-        border-bottom: 1px solid black;
-        padding-right: 4px;
-        font-family: 'line';
     }
-    .sum-amt:focus {
-        outline: none;
-    }
-    .paymt-container {
+
+    .paymt {
         display: flex;
-        margin-top: 10px;
+        margin-top: 15px;
         justify-content: space-between;
         font-weight: 400;
         font-size: 14px;
         color: #626262;
         margin-bottom: 30px;
+    }
+
+    .paymt-amt {
+        padding-right: 4px;
     }
     .paymt-label {
         margin-left: 68px;
@@ -136,9 +123,6 @@ const OptSecContainer = styled.div`
         font-size: 16px;
     }
 
-    .paymt-amt {
-        padding-right: 4px;
-    }
     .paymt-unit {
         margin-right: 20px;
         margin-left: 13px;
@@ -154,7 +138,7 @@ const OptSecContainer = styled.div`
         -moz-appearance: textfield;
     }
 `;
-const DutchPayDtl = styled.div`
+const CalDtl = styled.div`
     margin-top: 20px;
     margin-bottom: 50px;
     display: flex;
@@ -175,10 +159,19 @@ const DutchPayDtl = styled.div`
         position: relative;
     }
     .dt-value {
-        height: 26px;
+        font-size: 20px;
+        font-weight: 700;
+        width: 100px;
+        text-align: right;
+        height: 25px;
         line-height: 25px;
         padding-right: 4px;
+        font-family: 'line';
     }
+    .dt-value:focus {
+        outline: none;
+    }
+
     .main-dt-container {
         display: flex;
         font-weight: 700;
@@ -191,4 +184,4 @@ const DutchPayDtl = styled.div`
     }
 `;
 
-export default DutchPayUpdateOptSpl;
+export default CalDetailOptInp;
