+++ src/components/cartsharecal/update/CalUpdateOptInp.js
@@ -1,44 +1,30 @@
 /** @jsxImportSource @emotion/react */
 import styled from 'styled-components';
-import NameInfoCreate from './NameInfoCreate';
 
-const CartShareCalCreateOptSec = ({ cartShareCal, calcResponse }) => {
+import NameInfoUpdate from './NameInfoUpdate';
+
+const DutchPayUpdateOptInp = ({ cartShareCal, inpInput, onChangeInpInput, onChangeInpRmd }) => {
     return (
-        <OptSecContainer>
-            {calcResponse.cartShareCalDtlList?.map((info, index) => (
+        <OptInpContainer>
+            {cartShareCal.cartShareCalDtlList?.map((info, index) => (
                 <CalDtl key={info.mbrId}>
                     <CalDtlInfo>
-                        <NameInfoCreate info={info} />
-                        <div>
-                            <div className="main-cal">
-                                <div className="main-cal-amt">{info.calDtlAmt.toLocaleString()}</div>
-                                <div className="main-cal-unit">원</div>
-                            </div>
+                        <NameInfoUpdate info={info} />
+                        <div className="main-cal">
+                            <input
+                                className="main-cal-amt"
+                                onChange={e => onChangeInpInput(e, info.mbrId)}
+                                value={inpInput.dtlMap.get(info.mbrId).toLocaleString()}
+                            />
+                            <div className="main-cal-unit">원</div>
                         </div>
                     </CalDtlInfo>
-                    <div className="sub-cal-box">
-                        <div className="sub-cal">
-                            <div className="sub-cal-label">공동</div>
-                            <div className="sub-cal-amt">{info.commAmt.toLocaleString()}</div>
-                            <div className="sub-cal-unit">원</div>
-                        </div>
-                        <div className="sub-cal">
-                            <div className="sub-cal-label">개별</div>
-                            <div className="sub-cal-amt">{info.perAmt.toLocaleString()}</div>
-                            <div className="sub-cal-unit">원</div>
-                        </div>
-                        <div className="sub-cal">
-                            <div className="sub-cal-label sub-cal-last-label">배송비</div>
-                            <div className="sub-cal-amt">{info.shppCst.toLocaleString()}</div>
-                            <div className="sub-cal-unit">원</div>
-                        </div>
-                    </div>
                 </CalDtl>
             ))}
             <div className="rmd">
                 <div className="rmd-label">나머지</div>
                 <div className="rmd-value">
-                    <div className="rmd-amt">{calcResponse.calRmd}</div>
+                    <input className="rmd-amt" onChange={onChangeInpRmd} value={inpInput.calRmd.toLocaleString()} />
                     <div className="rmd-unit">원</div>
                 </div>
             </div>
@@ -46,7 +32,7 @@ const CartShareCalCreateOptSec = ({ cartShareCal, calcResponse }) => {
             <div className="sum">
                 <div className="sum-label">총 정산 금액</div>
                 <div className="sum-value">
-                    <div className="sum-amt">{calcResponse.calAmt.toLocaleString()}</div>
+                    <div className="sum-amt">{inpInput.calAmt.toLocaleString()}</div>
                     <div className="sum-unit">원</div>
                 </div>
             </div>
@@ -57,19 +43,19 @@ const CartShareCalCreateOptSec = ({ cartShareCal, calcResponse }) => {
                     <div className="paymt-unit">원</div>
                 </div>
             </div>
-        </OptSecContainer>
+        </OptInpContainer>
     );
 };
 
-const OptSecContainer = styled.div`
+const OptInpContainer = styled.div`
     .rmd {
         margin-top: 32px;
         display: flex;
         justify-content: space-between;
         font-weight: 400;
         font-size: 16px;
-        height: 15px;
-        line-height: 15px;
+        height: 20px;
+        line-height: 20px;
     }
     .rmd-label {
         padding-left: 46px;
@@ -81,7 +67,17 @@ const OptSecContainer = styled.div`
     }
 
     .rmd-amt {
-        margin-right: 3px;
+        width: 20px;
+        text-align: right;
+        border: none;
+        border-bottom: 1px solid black;
+        padding-right: 3px;
+        font-size: 16px;
+        font-family: 'line';
+        box-sizing: border-box;
+    }
+    .rmd-amt:focus {
+        outline: none;
     }
     .border {
         margin: auto;
@@ -106,15 +102,19 @@ const OptSecContainer = styled.div`
         font-weight: 700;
         font-size: 20px;
     }
-    .sum-amt {
-        font-weight: 700;
-        font-size: 24px;
-        padding-right: 4px;
-    }
     .sum-unit {
         margin-right: 20px;
         margin-left: 10px;
     }
+    .sum-amt {
+        font-weight: 700;
+        font-size: 24px;
+        border: none;
+        width: 100px;
+        text-align: right;
+        height: 24px;
+    }
+
     .paymt {
         display: flex;
         margin-top: 15px;
@@ -124,6 +124,10 @@ const OptSecContainer = styled.div`
         color: #626262;
         margin-bottom: 30px;
     }
+
+    .paymt-amt {
+        padding-right: 4px;
+    }
     .paymt-label {
         margin-left: 68px;
     }
@@ -131,47 +135,31 @@ const OptSecContainer = styled.div`
         display: flex;
         font-size: 16px;
     }
-    .paymt-amt {
-        padding-right: 4px;
-    }
+
     .paymt-unit {
         margin-right: 20px;
         margin-left: 14px;
     }
+    input::-webkit-outer-spin-button,
+    input::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+    }
+
+    /* Firefox */
+    input[type='number'] {
+        -moz-appearance: textfield;
+    }
 `;
+
 const CalDtl = styled.div`
     margin-top: 20px;
-    margin-bottom: 25px;
-
-    .sub-cal-box {
-        margin-bottom: 20px;
-    }
-    .sub-cal {
-        display: flex;
-        font-weight: 400;
-        font-size: 12px;
-        color: #888888;
-        height: 20px;
-        line-height: 20px;
-        justify-content: right;
-    }
-    .sub-cal-label {
-        width: 30px;
-        text-align: right;
-    }
-
-    .sub-cal-amt {
-        width: 60px;
-        text-align: right;
-    }
-    .sub-cal-unit {
-        margin-right: 14px;
-        margin-left: 3px;
-    }
+    margin-bottom: 50px;
 `;
 const CalDtlInfo = styled.div`
     display: flex;
     justify-content: space-between;
+
     .main-cal {
         display: flex;
         font-weight: 700;
@@ -182,6 +170,17 @@ const CalDtlInfo = styled.div`
     }
 
     .main-cal-amt {
+        border: none;
+        border-bottom: 1px solid black;
+        font-size: 20px;
+        font-weight: 700;
+        width: 100px;
+        text-align: right;
+        font-family: 'line';
+        box-sizing: border-box;
+    }
+    .main-cal-amt:focus {
+        outline: none;
     }
 
     .main-cal-unit {
@@ -190,4 +189,4 @@ const CalDtlInfo = styled.div`
     }
 `;
 
-export default CartShareCalCreateOptSec;
+export default DutchPayUpdateOptInp;
