+++ src/containers/cartsharecal/CartShareCalUpdateContainer.js
@@ -3,19 +3,25 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
 import { calCartShareCal, findCartShareCal, updateCartShareCal } from '../../api/cartsharecal/cartShareCal';
 
-import DutchPayUpdate from '../../components/cartsharecal/DutchPayUpdate';
+import CartShareCalUpdate from '../../components/cartsharecal/CartShareCalUpdate';
 
 const CartShareCalUpdateContainer = () => {
     const navigate = useNavigate();
     const { cartShareCalId } = useParams();
 
     const [cookies, setCookie] = useCookies(['mbrId']);
+    const [openOrdSheet, setOpenOrdSheet] = useState(false);
 
     const [opt, setOpt] = useState('');
-    const [cartShareCal, setCartShareCal] = useState({});
+    const [cartShareCal, setCartShareCal] = useState({
+        ttlPaymtAmt: 0,
+    });
 
     // for SECTION
-    const [calcResponse, setCalcResponse] = useState({});
+    const [calcResponse, setCalcResponse] = useState({
+        calAmt: 0,
+        calRmd: 0,
+    });
 
     // for SPLIT
     const [splInput, setSplInput] = useState({
@@ -40,6 +46,13 @@ const CartShareCalUpdateContainer = () => {
         navigate(`/cart-share-calculation/${cartShareCalId}`);
     };
 
+    const onClickOpenSheet = () => {
+        setOpenOrdSheet(true);
+    };
+    const onClickCloseSheet = () => {
+        setOpenOrdSheet(false);
+    };
+
     const onClickSave = () => {
         let request = {
             calRmd: 0,
@@ -52,8 +65,8 @@ const CartShareCalUpdateContainer = () => {
             case 'SECTION':
                 request.calRmd = calcResponse.calRmd;
                 request.calAmt = calcResponse.calAmt;
-                calcResponse.dutchPayDtlList.map((dtl, index) => {
-                    request.dutchPayDtlList.push({
+                calcResponse.cartShareCalDtlList.map((dtl, index) => {
+                    request.cartShareCalDtlList.push({
                         mbrId: dtl.mbrId,
                         calDtlAmt: dtl.calDtlAmt,
                         shppCst: dtl.shppCst,
@@ -63,6 +76,8 @@ const CartShareCalUpdateContainer = () => {
                 });
                 break;
             case 'SPLIT':
+                console.log(splInput);
+
                 request.calRmd = splInput.calRmd;
                 request.calAmt = splInput.calAmt;
                 request.calDtlAmt = splInput.calDtlAmt;
@@ -85,30 +100,30 @@ const CartShareCalUpdateContainer = () => {
     const calcSplit = n => {
         setSplInput({
             ...splInput,
-            dtAmt: n,
-            dtDtlAmt: parseInt(n / splInput.mbrNum),
-            dtRmd: n % splInput.mbrNum,
+            calAmt: n,
+            calDtlAmt: parseInt(n / splInput.mbrNum),
+            calRmd: n % splInput.mbrNum,
         });
     };
 
     const onChangeSplInput = e => {
-        calcSplit(Number(e.target.value));
+        calcSplit(Number(e.target.value.replace(/\D/g, '')));
     };
 
     const onChangeInpInput = (e, mbrId) => {
         setInpInput({
             ...inpInput,
-            dtlMap: new Map(inpInput.dtlMap.set(mbrId, Number(e.target.value))),
+            dtlMap: new Map(inpInput.dtlMap.set(mbrId, Number(e.target.value.replace(/\D/g, '')))),
         });
 
         let sum = 0;
         inpInput.dtlMap.forEach(v => {
             sum += v;
         });
-        sum += inpInput.dtRmd;
+        sum += inpInput.calRmd;
         setInpInput({
             ...inpInput,
-            dtAmt: sum,
+            calAmt: sum,
         });
     };
 
@@ -117,12 +132,11 @@ const CartShareCalUpdateContainer = () => {
         inpInput.dtlMap.forEach(v => {
             sum += v;
         });
-        sum += Number(e.target.value);
+        sum += Number(e.target.value.replace(/\D/g, ''));
         setInpInput({
             ...inpInput,
-            dtRmd: Number(e.target.value),
-
-            dtAmt: sum,
+            calRmd: Number(e.target.value.replace(/\D/g, '')),
+            calAmt: sum,
         });
     };
 
@@ -131,15 +145,15 @@ const CartShareCalUpdateContainer = () => {
             setCalcResponse(response.data.data);
         });
         findCartShareCal(cookies.mbrId, cartShareCalId).then(response => {
-            setOpt(response.data.data.dutchPayOptCd);
+            setOpt(response.data.data.calOptCd);
 
             setCartShareCal(response.data.data);
 
             setSplInput({
                 mbrNum: response.data.data.cartShareCalDtlList.length,
-                calAmt: response.data.data.calAmt,
+                calAmt: response.data.data.ttlPaymtAmt,
                 calDtlAmt: parseInt(response.data.data.ttlPaymtAmt / response.data.data.cartShareCalDtlList.length),
-                calRmd: response.data.data.paymtAmt % response.data.data.cartShareCalDtlList.length,
+                calRmd: response.data.data.ttlPaymtAmt % response.data.data.cartShareCalDtlList.length,
             });
             response.data.data.cartShareCalDtlList.map((info, index) => {
                 setInpInput({
@@ -150,31 +164,29 @@ const CartShareCalUpdateContainer = () => {
 
             if (response.data.data.calOptCd === 'SPLIT') {
                 setSplInput({
-                    ...splInput,
-                    calAmt: response.data.data.dutchPayAmt,
-                    calDtlAmt: response.data.data.dutchPayDtlList[0].dutchPayDtlAmt,
-                    calRmd: response.data.data.dutchPayRmd,
+                    mbrNum: response.data.data.cartShareCalDtlList.length,
+                    calAmt: response.data.data.calAmt,
+                    calDtlAmt: response.data.data.cartShareCalDtlList[0].calDtlAmt,
+                    calRmd: response.data.data.calRmd,
                 });
             }
             if (response.data.data.calOptCd === 'INPUT') {
-                console.log('HERE');
-                setInpInput({
-                    ...inpInput,
-                    calDtlAmt: response.data.data.calDtlAmt,
-                    calRmd: response.data.data.calRmd,
-                });
-
                 response.data.data.cartShareCalDtlList.map((info, index) => {
                     setInpInput({
                         ...inpInput,
-                        dtlMap: new Map(inpInput.dtlMap.set(info.mbrId, info.dutchPayDtlAmt)),
+                        dtlMap: new Map(inpInput.dtlMap.set(info.mbrId, info.calDtlAmt)),
                     });
+                });
+                setInpInput({
+                    ...inpInput,
+                    calAmt: response.data.data.calAmt,
+                    calRmd: response.data.data.calRmd,
                 });
             }
         });
     }, []);
     return (
-        <DutchPayUpdate
+        <CartShareCalUpdate
             onClickSave={onClickSave}
             cartShareCal={cartShareCal}
             opt={opt}
@@ -186,6 +198,9 @@ const CartShareCalUpdateContainer = () => {
             onChangeSplInput={onChangeSplInput}
             onChangeInpInput={onChangeInpInput}
             onChangeInpRmd={onChangeInpRmd}
+            openOrdSheet={openOrdSheet}
+            onClickOpenSheet={onClickOpenSheet}
+            onClickCloseSheet={onClickCloseSheet}
         />
     );
 };
